evttags: improve handling of NULL arguments to evt_tag_str
Earlier, evt_tag_str() simply returned NULL in case tag or value was NULL.
This is much less than ideal and can easily introduce memory leaks. With
NULL being returned evt_msg_create() simply assumes that it is at the
end of tags, discarding any further tags in the argument list.

This patch adds an assert on tag-name (as that's usually a literal) and
replaces NULL-values with a '(null)' string.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>